+++ BestRecipes/Resources/Assets.xcassets/Images/RecipeItem/BookmarkIcons/Contents.json
@@ -1,0 +1,6 @@
+{
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
