+++ BestRecipes/Resources/Assets.xcassets/Images/RecipeItem/authorPlaceholder.imageset/Contents.json
