+++ BestRecipes/Resources/Assets.xcassets/Images/RecipeItem/authorPlaceholder.imageset/authorPlaceholder.png
