+++ BestRecipes/Resources/Assets.xcassets/Images/RecipePhotoSample.imageset/Contents.json
